Move remaining per-call methods to dedicated util files
