+++ packages/client/src/rpc/modules/engine/util/generic.ts
@@ -1,19 +1,11 @@
 import { Block } from '@ethereumjs/block'
 import { Hardfork } from '@ethereumjs/common'
-import {
-  BIGINT_1,
-  bytesToHex,
-  bytesToUnprefixedHex,
-  equalsBytes,
-  hexToBytes,
-} from '@ethereumjs/util'
+import { BIGINT_1, bytesToHex, bytesToUnprefixedHex, equalsBytes } from '@ethereumjs/util'
 
 import { UNSUPPORTED_FORK } from '../../../error-code'
-import { type ChainCache, type PayloadStatusV1, Status } from '../types'
+import { type ChainCache } from '../types'
 
 import type { Chain } from '../../../../blockchain'
-import type { ExecutionPayloadBodyV1 } from '../types'
-import type { ExecutionPayload } from '@ethereumjs/block'
 import type { Common } from '@ethereumjs/common'
 
 /**
@@ -136,54 +128,6 @@ export const validateTerminalBlock = async (block: Block, chain: Chain): Promise
 
   const parentBlockTd = await chain.getTd(block.header.parentHash, block.header.number - BIGINT_1)
   return blockTd >= ttd && parentBlockTd < ttd
-}
-
-/**
- * Returns a block from a payload.
- * If errors, returns {@link PayloadStatusV1}
- */
-export const assembleBlock = async (
-  payload: ExecutionPayload,
-  chain: Chain,
-  chainCache: ChainCache
-): Promise<{ block?: Block; error?: PayloadStatusV1 }> => {
-  const { blockNumber, timestamp } = payload
-  const { config } = chain
-  const common = config.chainCommon.copy()
-
-  // This is a post merge block, so set its common accordingly
-  // Can't use setHardfork flag, as the transactions will need to be deserialized
-  // first before the header can be constucted with their roots
-  const ttd = common.hardforkTTD(Hardfork.Paris)
-  common.setHardforkBy({ blockNumber, td: ttd !== null ? ttd : undefined, timestamp })
-
-  try {
-    const block = await Block.fromExecutionPayload(payload, { common })
-    // TODO: validateData is also called in applyBlock while runBlock, may be it can be optimized
-    // by removing/skipping block data validation from there
-    await block.validateData()
-    return { block }
-  } catch (error) {
-    const validationError = `Error assembling block from payload: ${error}`
-    config.logger.error(validationError)
-    const latestValidHash = await validHash(hexToBytes(payload.parentHash), chain, chainCache)
-    const response = {
-      status: `${error}`.includes('Invalid blockHash') ? Status.INVALID_BLOCK_HASH : Status.INVALID,
-      latestValidHash,
-      validationError,
-    }
-    return { error: response }
-  }
-}
-
-export const getPayloadBody = (block: Block): ExecutionPayloadBodyV1 => {
-  const transactions = block.transactions.map((tx) => bytesToHex(tx.serialize()))
-  const withdrawals = block.withdrawals?.map((wt) => wt.toJSON()) ?? null
-
-  return {
-    transactions,
-    withdrawals,
-  }
 }
 
 export function validateHardforkRange(
