+++ packages/client/src/rpc/modules/engine/util/index.ts
@@ -1,4 +1,5 @@
 export * from './forkchoiceUpdated'
 export * from './generic'
 export * from './getPayload'
+export * from './getPayloadBody'
 export * from './newPayload'
